feat: Add git alias for creating fixup commits
The alias will use the fzf fuzzy finder to select the commit to fixup.
The sha of the selected commit will be inserted into `git commit
--fixup=...`.

Before, it seemed common to perform this in multiple steps: find the
commit to fixup, copy the sha, write the commit statement, and finally
paste in the sha. These operations are combined into this alias.